moved all fixtures into test files
