+++ tests/functional/simple_snapshot/fixtures.py
@@ -1,7 +1,3 @@
-import pytest
-from dbt.tests.fixtures.project import write_project_files
-
-
 snapshots_select__snapshot_sql = """
 {% snapshot snapshot_castillo %}
 
@@ -359,92 +355,3 @@ snapshots_check_col_noconfig__snapshot_sql = """
     select * from {{target.database}}.{{schema}}.seed
 {% endsnapshot %}
 """
-
-
-@pytest.fixture(scope="class")
-def snapshots_select():
-    return {
-        "snapshot.sql": snapshots_pg__snapshot_sql,
-        "snapshot_select.sql": snapshots_select__snapshot_sql,
-    }
-
-
-@pytest.fixture(scope="class")
-def snapshots_pg_custom():
-    return {"snapshot.sql": snapshots_pg_custom__snapshot_sql}
-
-
-@pytest.fixture(scope="class")
-def macros_custom_snapshot():
-    return {
-        "test_no_overlaps.sql": macros__test_no_overlaps_sql,
-        "custom.sql": macros_custom_snapshot__custom_sql,
-    }
-
-
-@pytest.fixture(scope="class")
-def models():
-    return {
-        "schema.yml": models__schema_yml,
-        "ref_snapshot.sql": models__ref_snapshot_sql,
-    }
-
-
-@pytest.fixture(scope="class")
-def macros():
-    return {"test_no_overlaps.sql": macros__test_no_overlaps_sql}
-
-
-@pytest.fixture(scope="class")
-def snapshots_select_noconfig():
-    return {"snapshot.sql": snapshots_select_noconfig__snapshot_sql}
-
-
-@pytest.fixture(scope="class")
-def seeds():
-    return {"seed_newcol.csv": seeds__seed_newcol_csv, "seed.csv": seeds__seed_csv}
-
-
-@pytest.fixture(scope="class")
-def snapshots_pg_custom_namespaced():
-    return {"snapshot.sql": snapshots_pg_custom_namespaced__snapshot_sql}
-
-
-@pytest.fixture(scope="class")
-def snapshots_pg():
-    return {"snapshot.sql": snapshots_pg__snapshot_sql}
-
-
-@pytest.fixture(scope="class")
-def models_slow():
-    return {"gen.sql": models_slow__gen_sql}
-
-
-@pytest.fixture(scope="class")
-def project_files(
-    project_root,
-    snapshots_select,
-    snapshots_pg_custom,
-    macros_custom_snapshot,
-    models,
-    macros,
-    snapshots_select_noconfig,
-    seeds,
-    snapshots_pg_custom_namespaced,
-    snapshots_pg,
-    models_slow,
-):
-    write_project_files(project_root, "snapshots-select", snapshots_select)
-    write_project_files(project_root, "snapshots-pg-custom", snapshots_pg_custom)
-    write_project_files(project_root, "macros-custom-snapshot", macros_custom_snapshot)
-    write_project_files(project_root, "models", models)
-    write_project_files(project_root, "macros", macros)
-    write_project_files(project_root, "snapshots-select-noconfig", snapshots_select_noconfig)
-    #     )
-    write_project_files(project_root, "seeds", seeds)
-    write_project_files(
-        project_root, "snapshots-pg-custom-namespaced", snapshots_pg_custom_namespaced
-    )
-    write_project_files(project_root, "snapshots-pg", snapshots_pg)
-
-    write_project_files(project_root, "models-slow", models_slow)
