+++ tests/functional/simple_snapshot/test_changing_strategy_snapshot.py
@@ -1,8 +1,6 @@
 import pytest
 from dbt.tests.util import run_dbt
-from tests.functional.simple_snapshot.fixtures import (  # noqa: F401
-    models_slow,
-)
+from tests.functional.simple_snapshot.fixtures import models_slow__gen_sql
 
 
 test_snapshots_changing_strategy__test_snapshot_sql = """
@@ -101,8 +99,8 @@ snapshots_changing_strategy__snapshot_sql = """
 
 
 @pytest.fixture(scope="class")
-def models(models_slow):  # noqa: F811
-    return models_slow
+def models():
+    return {"gen.sql": models_slow__gen_sql}
 
 
 @pytest.fixture(scope="class")
