+++ tests/functional/simple_snapshot/test_cross_schema_snapshot.py
@@ -1,15 +1,33 @@
 import os
 import pytest
 from dbt.tests.util import run_dbt
-from tests.functional.simple_snapshot.fixtures import models, macros, snapshots_pg  # noqa: F401
+from tests.functional.simple_snapshot.fixtures import (
+    models__schema_yml,
+    models__ref_snapshot_sql,
+    macros__test_no_overlaps_sql,
+    snapshots_pg__snapshot_sql,
+)
 
 
 NUM_SNAPSHOT_MODELS = 1
 
 
 @pytest.fixture(scope="class")
-def snapshots(snapshots_pg):  # noqa: F811
-    return snapshots_pg
+def snapshots():
+    return {"snapshot.sql": snapshots_pg__snapshot_sql}
+
+
+@pytest.fixture(scope="class")
+def models():
+    return {
+        "schema.yml": models__schema_yml,
+        "ref_snapshot.sql": models__ref_snapshot_sql,
+    }
+
+
+@pytest.fixture(scope="class")
+def macros():
+    return {"test_no_overlaps.sql": macros__test_no_overlaps_sql}
 
 
 def test_cross_schema_snapshot(project):
