+++ tests/functional/simple_snapshot/test_missing_strategy_snapshot.py
@@ -1,9 +1,10 @@
 import pytest
 from dbt.tests.util import run_dbt
 from dbt.exceptions import ParsingException
-from tests.functional.simple_snapshot.fixtures import (  # noqa: F401
-    models,
-    macros,
+from tests.functional.simple_snapshot.fixtures import (
+    models__schema_yml,
+    models__ref_snapshot_sql,
+    macros__test_no_overlaps_sql,
 )
 
 snapshots_invalid__snapshot_sql = """
@@ -26,6 +27,19 @@ snapshots_invalid__snapshot_sql = """
 @pytest.fixture(scope="class")
 def snapshots():
     return {"snapshot.sql": snapshots_invalid__snapshot_sql}
+
+
+@pytest.fixture(scope="class")
+def models():
+    return {
+        "schema.yml": models__schema_yml,
+        "ref_snapshot.sql": models__ref_snapshot_sql,
+    }
+
+
+@pytest.fixture(scope="class")
+def macros():
+    return {"test_no_overlaps.sql": macros__test_no_overlaps_sql}
 
 
 def test_missing_strategy(project):
